# perf: move collector queue metrics to be outside of the collect loop