## Which problem is this PR solving?

Previously, both collector_incoming_queue and collector_peer_queue lengths were being recorded inside the collect loop. Since these are channels, we can safely move the length recording logic out of the loop. This change frees up the collect loop to focus more on ingesting and processing span data, improving performance.

Additionally, this PR reduces the frequency of metric recording for these queues. These metrics are currently recorded using histograms, which are relatively expensive in terms of allocations. Prior to this change, they were being recorded at least once per ingested span or every `SendTicker` — a frequency that’s unnecessarily high and likely contributing to excessive memory usage.

By moving the metric recording to a separate goroutine and throttling updates to once per 100ms, we significantly reduce allocations along the hot path.

## Short description of the changes

- record queue metrics in a separate goroutine with a 100ms ticker

## Benchmark Result
```
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/refinery/collect
cpu: Apple M2 Max
                                                     │ old-metric.text │           new-metric.text           │
                                                     │     sec/op      │    sec/op     vs base               │
CollectorWithSamplers/deterministic/small_traces-12        90.71µ ± 6%   86.34µ ±  5%  -4.83% (p=0.003 n=10)
CollectorWithSamplers/deterministic/medium_traces-12       8.694m ± 5%   8.237m ±  2%  -5.26% (p=0.001 n=10)
CollectorWithSamplers/deterministic/large_traces-12        899.2m ± 6%   886.8m ± 10%       ~ (p=0.579 n=10)
CollectorWithSamplers/dynamic/small_traces-12              91.40µ ± 3%   87.74µ ±  3%  -4.00% (p=0.001 n=10)
CollectorWithSamplers/dynamic/medium_traces-12             8.753m ± 3%   8.353m ±  3%  -4.57% (p=0.005 n=10)
CollectorWithSamplers/dynamic/large_traces-12             1009.7m ± 3%   947.8m ±  7%  -6.13% (p=0.019 n=10)
CollectorWithSamplers/emadynamic/small_traces-12           88.56µ ± 2%   87.08µ ±  2%       ~ (p=0.052 n=10)
CollectorWithSamplers/emadynamic/medium_traces-12          8.476m ± 2%   8.189m ±  3%  -3.39% (p=0.002 n=10)
CollectorWithSamplers/emadynamic/large_traces-12           958.0m ± 3%   927.6m ±  2%  -3.17% (p=0.005 n=10)
CollectorWithSamplers/rulesbased/small_traces-12           82.97µ ± 2%   80.86µ ±  2%  -2.55% (p=0.002 n=10)
CollectorWithSamplers/rulesbased/medium_traces-12          8.186m ± 2%   7.736m ±  3%  -5.49% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/large_traces-12           891.8m ± 4%   864.2m ±  2%  -3.09% (p=0.000 n=10)
geomean                                                    8.908m        8.569m        -3.80%

                                                     │ old-metric.text │           new-metric.text            │
                                                     │      B/op       │     B/op      vs base                │
CollectorWithSamplers/deterministic/small_traces-12       4.646Ki ± 0%   3.082Ki ± 0%  -33.67% (p=0.000 n=10)
CollectorWithSamplers/deterministic/medium_traces-12      461.0Ki ± 0%   304.7Ki ± 0%  -33.91% (p=0.000 n=10)
CollectorWithSamplers/deterministic/large_traces-12       58.18Mi ± 0%   42.92Mi ± 0%  -26.23% (p=0.000 n=10)
CollectorWithSamplers/dynamic/small_traces-12             5.689Ki ± 0%   4.122Ki ± 0%  -27.54% (p=0.000 n=10)
CollectorWithSamplers/dynamic/medium_traces-12            462.1Ki ± 0%   305.8Ki ± 0%  -33.83% (p=0.000 n=10)
CollectorWithSamplers/dynamic/large_traces-12             58.19Mi ± 0%   42.92Mi ± 0%  -26.24% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/small_traces-12          5.767Ki ± 0%   4.200Ki ± 0%  -27.16% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/medium_traces-12         462.1Ki ± 0%   305.8Ki ± 0%  -33.82% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/large_traces-12          58.19Mi ± 0%   42.92Mi ± 0%  -26.23% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/small_traces-12          5.428Ki ± 0%   3.867Ki ± 0%  -28.75% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/medium_traces-12         461.8Ki ± 0%   305.5Ki ± 0%  -33.85% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/large_traces-12          58.19Mi ± 0%   42.92Mi ± 0%  -26.23% (p=0.000 n=10)
geomean                                                   528.4Ki        370.6Ki       -29.88%

                                                     │ old-metric.text │           new-metric.text            │
                                                     │    allocs/op    │  allocs/op    vs base                │
CollectorWithSamplers/deterministic/small_traces-12        226.00 ± 0%    26.00 ±  0%  -88.50% (p=0.000 n=10)
CollectorWithSamplers/deterministic/medium_traces-12     20052.00 ± 0%    48.00 ±  2%  -99.76% (p=0.000 n=10)
CollectorWithSamplers/deterministic/large_traces-12     2001.614k ± 0%   1.288k ± 10%  -99.94% (p=0.000 n=10)
CollectorWithSamplers/dynamic/small_traces-12              256.00 ± 0%    56.00 ±  0%  -78.12% (p=0.000 n=10)
CollectorWithSamplers/dynamic/medium_traces-12           20085.00 ± 0%    81.00 ±  0%  -99.60% (p=0.000 n=10)
CollectorWithSamplers/dynamic/large_traces-12           2001.760k ± 0%   1.320k ±  9%  -99.93% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/small_traces-12           258.00 ± 0%    58.00 ±  0%  -77.52% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/medium_traces-12        20087.00 ± 0%    83.00 ±  0%  -99.59% (p=0.000 n=10)
CollectorWithSamplers/emadynamic/large_traces-12        2001.655k ± 0%   1.302k ±  3%  -99.93% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/small_traces-12           235.00 ± 0%    35.00 ±  0%  -85.11% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/medium_traces-12        20062.00 ± 0%    58.00 ±  2%  -99.71% (p=0.000 n=10)
CollectorWithSamplers/rulesbased/large_traces-12        2001.735k ± 0%   1.389k ±  3%  -99.93% (p=0.000 n=10)
geomean                                                    21.38k         153.4        -99.28%
```